+++ streamraindatalayer/src/main/java/com/tsi2/streamrain/dao/implementations/StreamRainMySQLMainDAO.java
@@ -24,7 +24,7 @@ public class StreamRainMySQLMainDAO implements IDAOService {
 		dbSession.delete(object);
 	}
 
-	public <T> T get(final Class<T> type, final Long id, final String tenantID) {
+	public <T> T get(final Class<T> type, final Integer id, final String tenantID) {
 		Session dbSession = DBHibernateUtil.getSessionFactoryMain();
 		return (T) dbSession.get(type, id);
 	}
